refactored common functions to lib/common.js
