+++ lib/ResourceGenerator.js
@@ -1,6 +1,7 @@
-const chalk = require('chalk')
 const fse = require('fs-extra')
 const Resource = require(`${__dirname}/Resource`)
+
+require(`${__dirname}/common.js`)()
 
 class ResourceGenerator {
   static createResourcesFromPath(path) {
@@ -125,28 +126,6 @@ class ResourceGenerator {
     replaceString(modelTarget, /\{\{model_attributes\}\}/g, modelReplacement)
   }
   /* == MISC METHODS END == */
-}
-
-let capitalize = ([first,...rest]) => first.toUpperCase() + rest.join('')
-
-let replaceString = (file, target, value) => {
-  try {
-    let data = fse.readFileSync(file, 'utf8')
-    let result = data.replace(target, value)
-    fse.writeFileSync(file, result, 'utf8')
-  } catch (err) {
-    throw err
-  }
-}
-
-let copyWithDisplay = (target, endpoint) => {
-  fse.copySync(target, endpoint)
-  console.log(chalk.bgGreen(chalk.bold(' CREATE ')), endpoint)
-}
-
-let createDirectoryWithDisplay = (path) => {
-  fse.ensureDirSync(path)
-  console.log(chalk.bgGreen(chalk.bold(' CREATE ')), path)
 }
 
 module.exports = ResourceGenerator
